% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tbl_na}
\alias{check_tbl_na}
\title{check for NA values in a tibble}
\usage{
check_tbl_na(tbl)
}
\arguments{
\item{tbl}{A \code{tibble}}
}
\value{
A logical value.
}
\description{
check for NA values in a tibble
}
\examples{
demotbl = tibble::tibble(x = c(1,2,3,NA,1),
                         y = c(NA,NA,1:3),
                         z = 1:5)
demotbl
check_tbl_na(demotbl)

}
