% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{(partial) correlation test}
\usage{
cor_test(x, y, z = NULL, level = 0.05)
}
\arguments{
\item{x}{A numeric vector representing the first variable.}

\item{y}{A numeric vector representing the second variable.}

\item{z}{An optional numeric vector or matrix of control variables. If provided, partial correlation is computed.}

\item{level}{(optional) Significance level. Default is 0.05.}
}
\value{
A numeric vector
}
\description{
(partial) correlation test
}
\examples{
gzma = sf::read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
cor_test(gzma$PS_Score,gzma$EL_Score)

}
