% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_var.R
\name{dummy_tbl}
\alias{dummy_tbl}
\title{transforming a category tibble into the corresponding dummy variable tibble}
\usage{
dummy_tbl(tbl)
}
\arguments{
\item{tbl}{A \code{tibble} or \code{data.frame}.}
}
\value{
A \code{tibble}
}
\description{
transforming a category tibble into the corresponding dummy variable tibble
}
\examples{
a = tibble::tibble(x = 1:3,y = 4:6)
dummy_tbl(a)

}
