% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_var.R
\name{dummy_vec}
\alias{dummy_vec}
\title{transforming a categorical variable into dummy variables}
\usage{
dummy_vec(x)
}
\arguments{
\item{x}{An integer vector or can be converted into an integer vector.}
}
\value{
A matrix.
}
\description{
transforming a categorical variable into dummy variables
}
\examples{
dummy_vec(c(1,1,3,2,4,6))

}
