% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_test.R
\name{geodetector_q}
\alias{geodetector_q}
\title{only geodetector q-value}
\usage{
geodetector_q(y, hs)
}
\arguments{
\item{y}{Dependent variable}

\item{hs}{Independent variable}
}
\value{
A numeric value
}
\description{
only geodetector q-value
}
\examples{
geodetector_q(y = 1:7, hs = c('x',rep('y',3),rep('z',3)))

}
