% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{hclustgeo_disc}
\alias{hclustgeo_disc}
\title{hierarchical clustering with spatial soft constraints}
\usage{
hclustgeo_disc(sfj, n, alpha = 0.5, scale = TRUE, wt = NULL, ...)
}
\arguments{
\item{sfj}{An \code{sf} object.}

\item{n}{The number of hierarchical clustering classes, which can be a numeric value or vector.}

\item{alpha}{(optional) A real value between 0 and 1. This mixing parameter gives the relative
importance of "feature"  space and constraint" space.  Default is \code{0.5}.}

\item{scale}{(optional) Whether to scaled the dissimilarities matrix, default is \code{TRUE}.}

\item{wt}{(optional) Vector with the weights of the observations. By default, \code{wt} is \code{NULL}.}

\item{...}{(optional) Other arguments passed to \code{stats::dist()}.}
}
\value{
A vector with group memberships if n are scalar, otherwise a matrix with group memberships
is returned where each column corresponds to the elements of n, respectively.
}
\description{
hierarchical clustering with spatial soft constraints
}
\examples{
gzma = sf::read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
gzma$group = hclustgeo_disc(gzma,5, alpha = 0.75)
plot(gzma["group"])

}
