% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_vector}
\alias{normalize_vector}
\title{normalization}
\usage{
normalize_vector(x, to_left = 0, to_right = 1)
}
\arguments{
\item{x}{A continuous numeric vector.}

\item{to_left}{(optional) Specified minimum. Default is \code{0}.}

\item{to_right}{(optional) Specified maximum. Default is \code{1}.}
}
\value{
A continuous vector which has normalized.
}
\description{
normalization
}
\examples{
normalize_vector(c(-5,1,5,0.01,0.99))

}
