% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{rm_lineartrend}
\alias{rm_lineartrend}
\title{remove variable linear trend based on covariate}
\usage{
rm_lineartrend(formula, data, method = c("cpp", "r"))
}
\arguments{
\item{formula}{A formula.}

\item{data}{The observation data.}

\item{method}{(optional) The method for using, which can be chosen as either \code{cpp}
or \code{r}. Default is \code{cpp}.}
}
\value{
A numeric vector.
}
\description{
remove variable linear trend based on covariate
}
\examples{
gzma = sf::read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
rm_lineartrend(PS_Score ~ ., gzma)
rm_lineartrend(PS_Score ~ ., gzma, method = "r")

}
