% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{sf_coordinates}
\alias{sf_coordinates}
\title{extract locations}
\usage{
sf_coordinates(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object or can be converted to \code{sf} by \code{sf::st_as_sf()}.}
}
\value{
A matrix.
}
\description{
Extract locations of sf objects.
}
\examples{
pts = sf::read_sf(system.file('extdata/pts.gpkg',package = 'sdsfun'))
sf_coordinates(pts)

}
