% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{sf_distance_matrix}
\alias{sf_distance_matrix}
\title{generates distance matrix}
\usage{
sf_distance_matrix(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object.}
}
\value{
A matix.
}
\description{
Generates distance matrix) for sf object
}
\examples{
library(sf)
pts = read_sf(system.file('extdata/pts.gpkg',package = 'sdsfun'))
pts_distm = sf_distance_matrix(pts)
pts_distm[1:5,1:5]

}
