% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{sf_geometry_name}
\alias{sf_geometry_name}
\title{sf object geometry column name}
\usage{
sf_geometry_name(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object.}
}
\value{
A character.
}
\description{
Get the geometry column name of an sf object
}
\examples{
library(sf)
gzma = read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
sf_geometry_name(gzma)

}
