% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{sf_geometry_type}
\alias{sf_geometry_type}
\title{sf object geometry type}
\usage{
sf_geometry_type(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object.}
}
\value{
A lowercase character vector
}
\description{
Get the geometry type of an sf object
}
\examples{
library(sf)
snnu = read_sf(system.file('extdata/snnu.geojson',package = 'sdsfun'))
sf_geometry_type(snnu)

}
