% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{sf_utm_proj_wgs84}
\alias{sf_utm_proj_wgs84}
\title{generates wgs84 utm projection epsg coding character}
\usage{
sf_utm_proj_wgs84(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object.}
}
\value{
A character.
}
\description{
Generates a utm projection epsg coding character corresponding to an \code{sfj} object
under the WGS84 spatial reference.
}
\details{
For more details, please refer to https://zhuanlan.zhihu.com/p/670055831
}
\examples{
library(sf)
snnu = read_sf(system.file('extdata/snnu.geojson',package = 'sdsfun'))
sf_utm_proj_wgs84(snnu)

}
