% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvar.R
\name{spade_psd}
\alias{spade_psd}
\title{only spade power of spatial determinant}
\usage{
spade_psd(y, hs, wt)
}
\arguments{
\item{y}{Dependent variable}

\item{hs}{Independent variable}

\item{wt}{Spatial weight matrix}
}
\value{
A numeric value
}
\description{
only spade power of spatial determinant
}
\examples{
gzma = sf::read_sf(system.file('extdata/gzma.gpkg',package = 'sdsfun'))
wt1 = inverse_distance_swm(gzma)
spade_psd(y = gzma$PS_Score,
          hs = discretize_vector(gzma$PS_Score,5),
          wt = wt1)

}
