% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_test.R
\name{ssh_test}
\alias{ssh_test}
\title{spatial stratified heterogeneity test}
\usage{
ssh_test(y, hs)
}
\arguments{
\item{y}{Variable Y, continuous numeric vector.}

\item{hs}{Spatial stratification or classification of each explanatory variable.
\code{factor}, \code{character}, \code{integer} or \code{data.frame}, \code{tibble} and \code{sf} object.}
}
\value{
A \code{tibble}
}
\description{
spatial stratified heterogeneity test
}
\examples{
ssh_test(y = 1:7, hs = c('x',rep('y',3),rep('z',3)))

}
