% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_all2int}
\alias{tbl_all2int}
\title{convert discrete variables in a tibble to integers}
\usage{
tbl_all2int(tbl)
}
\arguments{
\item{tbl}{A \code{tibble},\code{data.frame} or \code{sf} object.}
}
\value{
A converted \code{tibble},\code{data.frame} or \code{sf} object.
}
\description{
convert discrete variables in a tibble to integers
}
\examples{
demotbl = tibble::tibble(x = c(1,2,3,3,1),
                         y = letters[1:5],
                         z = c(1L,1L,2L,2L,3L),
                         m = factor(letters[1:5],levels = letters[5:1]))
tbl_all2int(demotbl)

}
