% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_test.R
\name{ssh_test}
\alias{ssh_test}
\title{test explanatory power of spatial stratified heterogeneity}
\usage{
ssh_test(y, hs)
}
\arguments{
\item{y}{Variable Y, continuous numeric vector.}

\item{hs}{Spatial partitioning or classification of each explanatory variable.
\code{factor}, \code{character}, \code{integer} or \code{data.frame}, \code{tibble}.}
}
\value{
A \code{tibble}
}
\description{
Spatial stratified heterogeneity test based on geographical detector q value.
}
\note{
This is a \verb{C++} implementation of the \code{factor_detector} function in \code{gdverse} package.
}
\examples{
ssh_test(y = 1:7, hs = c('x',rep('y',3),rep('z',3)))

}
